split db access logic from ui
